Actualizacion general incluyendo temas pendientes de la planilla de pago

+++ src/Planillas/CoreBundle/Controller/CDeudasController.php
@@ -5,7 +5,6 @@ namespace Planillas\CoreBundle\Controller;
 use Planillas\CoreBundle\Form\Type\BuscarDeudasType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-
 use Planillas\CoreBundle\Entity\CDeudas;
 use Planillas\CoreBundle\Form\Type\CDeudasType;
 
@@ -13,15 +12,13 @@ use Planillas\CoreBundle\Form\Type\CDeudasType;
  * CDeudas controller.
  *
  */
-class CDeudasController extends Controller
-{
+class CDeudasController extends Controller {
 
     /**
      * Lists all CDeudas entities.
      *
      */
-    public function indexAction()
-    {
+    public function indexAction() {
         $em = $this->getDoctrine()->getManager();
         $request = $this->getRequest();
         $aDatos = array();
@@ -36,14 +33,14 @@ class CDeudasController extends Controller
         $paginator = $this->get('knp_paginator');
         $session = $this->get('session')->set('filtros', array()); //hay que meter la busqueda en la sesion
         $pagination = $paginator->paginate(
-            $result, $this->get('request')->query->get('page', 1), 10
+                $result, $this->get('request')->query->get('page', 1), 10
         );
         $entity = new CDeudas();
         $form_new = $this->createCreateForm($entity);
         return $this->render('PlanillasCoreBundle:CDeudas:index.html.twig', array(
-            'pagination' => $pagination,
-            'form_buscar' => $form->createView(),
-            'form' => $form_new->createView()
+                    'pagination' => $pagination,
+                    'form_buscar' => $form->createView(),
+                    'form' => $form_new->createView()
         ));
     }
 
@@ -51,8 +48,7 @@ class CDeudasController extends Controller
      * Creates a new CDeudas entity.
      *
      */
-    public function createAction(Request $request)
-    {
+    public function createAction(Request $request) {
         $entity = new CDeudas();
         $em = $this->getDoctrine()->getManager();
         $form = new CDeudasType();
@@ -64,13 +60,11 @@ class CDeudasController extends Controller
             $entity = $em->getRepository('PlanillasCoreBundle:CDeudas')->find($data['id']);
             if (!$entity) {
                 throw $this->createNotFoundException('Unable to find CDeudas entity.');
-
             } else {
 
 
                 $form = $this->createEditForm($entity);
             }
-
         } else {
             //$entity->setMontoRestante(0);
             $form = $this->createCreateForm($entity);
@@ -99,8 +93,7 @@ class CDeudasController extends Controller
      *
      * @return \Symfony\Component\Form\Form The form
      */
-    private function createCreateForm(CDeudas $entity)
-    {
+    private function createCreateForm(CDeudas $entity) {
         $form = $this->createForm(new CDeudasType(), $entity, array(
             'action' => $this->generateUrl('cdeudas_create'),
             'method' => 'POST',
@@ -115,14 +108,13 @@ class CDeudasController extends Controller
      * Displays a form to create a new CDeudas entity.
      *
      */
-    public function newAction()
-    {
+    public function newAction() {
         $entity = new CDeudas();
         $form = $this->createCreateForm($entity);
 
         return $this->render('PlanillasCoreBundle:CDeudas:new.html.twig', array(
-            'entity' => $entity,
-            'form' => $form->createView(),
+                    'entity' => $entity,
+                    'form' => $form->createView(),
         ));
     }
 
@@ -130,8 +122,7 @@ class CDeudasController extends Controller
      * Finds and displays a CDeudas entity.
      *
      */
-    public function showAction($id)
-    {
+    public function showAction($id) {
         $em = $this->getDoctrine()->getManager();
 
         $entity = $em->getRepository('PlanillasCoreBundle:CDeudas')->find($id);
@@ -143,16 +134,15 @@ class CDeudasController extends Controller
         $deleteForm = $this->createDeleteForm($id);
 
         return $this->render('PlanillasCoreBundle:CDeudas:show.html.twig', array(
-            'entity' => $entity,
-            'delete_form' => $deleteForm->createView(),));
+                    'entity' => $entity,
+                    'delete_form' => $deleteForm->createView(),));
     }
 
     /**
      * Displays a form to edit an existing CDeudas entity.
      *
      */
-    public function editAction($id)
-    {
+    public function editAction($id) {
         $em = $this->getDoctrine()->getManager();
 
         $entity = $em->getRepository('PlanillasCoreBundle:CDeudas')->find($id);
@@ -165,9 +155,9 @@ class CDeudasController extends Controller
         $deleteForm = $this->createDeleteForm($id);
 
         return $this->render('PlanillasCoreBundle:CDeudas:edit.html.twig', array(
-            'entity' => $entity,
-            'edit_form' => $editForm->createView(),
-            'delete_form' => $deleteForm->createView(),
+                    'entity' => $entity,
+                    'edit_form' => $editForm->createView(),
+                    'delete_form' => $deleteForm->createView(),
         ));
     }
 
@@ -178,8 +168,7 @@ class CDeudasController extends Controller
      *
      * @return \Symfony\Component\Form\Form The form
      */
-    private function createEditForm(CDeudas $entity)
-    {
+    private function createEditForm(CDeudas $entity) {
         $form = $this->createForm(new CDeudasType(), $entity, array(
             'action' => $this->generateUrl('cdeudas_update', array('id' => $entity->getId())),
             'method' => 'POST',
@@ -194,8 +183,7 @@ class CDeudasController extends Controller
      * Edits an existing CDeudas entity.
      *
      */
-    public function updateAction(Request $request, $id)
-    {
+    public function updateAction(Request $request, $id) {
         $em = $this->getDoctrine()->getManager();
 
         $entity = $em->getRepository('PlanillasCoreBundle:CDeudas')->find($id);
@@ -215,9 +203,9 @@ class CDeudasController extends Controller
         }
 
         return $this->render('PlanillasCoreBundle:CDeudas:edit.html.twig', array(
-            'entity' => $entity,
-            'edit_form' => $editForm->createView(),
-            'delete_form' => $deleteForm->createView(),
+                    'entity' => $entity,
+                    'edit_form' => $editForm->createView(),
+                    'delete_form' => $deleteForm->createView(),
         ));
     }
 
@@ -225,8 +213,7 @@ class CDeudasController extends Controller
      * Deletes a CDeudas entity.
      *
      */
-    public function deleteAction(Request $request, $id)
-    {
+    public function deleteAction(Request $request, $id) {
 
         try {
 
@@ -236,16 +223,17 @@ class CDeudasController extends Controller
             if (!$entity) {
                 throw $this->createNotFoundException('Unable to find CDeudas entity.');
             }
-
+            if ($entity->getPlanilla() != null) {
+                $this->get('session')->getFlashBag()->add('danger', 'No puede eliminar la entidad porque ya que está asociada a una planilla de efectivo');
+                return $this->redirect($this->generateUrl('cdeudas'));
+            }
             $em->remove($entity);
             $em->flush();
             $this->get('session')->getFlashBag()->add('info', 'Los datos han sido eliminados correctamente');
-
         } catch (Exception $e) {
             $this->get('session')->getFlashBag()->add('info', 'No se pudieron eliminar los datos');
         }
         return $this->redirect($this->generateUrl('cdeudas'));
-
     }
 
     /**
@@ -255,19 +243,16 @@ class CDeudasController extends Controller
      *
      * @return \Symfony\Component\Form\Form The form
      */
-    private function createDeleteForm($id)
-    {
+    private function createDeleteForm($id) {
         return $this->createFormBuilder()
-            ->setAction($this->generateUrl('cdeudas_delete', array('id' => $id)))
-            ->setMethod('DELETE')
-            ->add('submit', 'submit', array('label' => 'Delete'))
-            ->getForm();
+                        ->setAction($this->generateUrl('cdeudas_delete', array('id' => $id)))
+                        ->setMethod('DELETE')
+                        ->add('submit', 'submit', array('label' => 'Delete'))
+                        ->getForm();
     }
 
-    public function editajaxAction(Request $request)
-    {
+    public function editajaxAction(Request $request) {
         $id = $request->get('id');
-        //print_r($id);exit;
         $em = $this->getDoctrine()->getManager();
         $entity = $em->getRepository('PlanillasCoreBundle:CDeudas')->find($id);
         $response = array("success" => true);
@@ -275,7 +260,14 @@ class CDeudasController extends Controller
             $response['success'] = false;
             return new \Symfony\Component\HttpFoundation\Response(json_encode($response));
         }
+        if ($entity->getPlanilla() != null) {
+            $response['success'] = false;
+            $response['mensaje'] = 'No puede editar la entidad ya que está asociada a una planilla de efectivo.';
+            return new \Symfony\Component\HttpFoundation\Response(json_encode($response));
+        }
+
         $response['data'] = $entity->getJson();
         return new \Symfony\Component\HttpFoundation\Response(json_encode($response));
     }
+
 }
